Moved subcommand into another file
